+++ CHANGELOG.md
@@ -9,9 +9,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added injection of Oauth token on all service methods when using DsPresentClient
 - IIIF metods not supported by service2service Oauth since no java client is generated for these methods. Also not intended to be called by other modules.
 
+### Added
+- ds-present-transformation-errors log which logs only transformation errors.
+
+### Changed
+- Bumped multiple OpenAPI dependency versions
+- Changed the response format for the JSON response in /records to include information on failed transformations.
+
+### Fixed
+- Fixed /api-docs wrongly showing petstore example API spec
+- Fixed genre unittests failing wrongly
+
+## [2.2.6](https://github.com/kb-dk/ds-present/releases/tag/ds-present-2.2.6) 2025-01-27
+### Changed
+- Bumped SwaggerUI dependency to v5.18.2
+
 ### Fixed
 - Fixed inclusion of the same dependencies from multiple sources.
 - Fixed that not all records did get a fallback genre.
+- Solr 1.7.9: Fixed field genre_facet not being added as a facet field.
+
 
 ## [2.2.5](https://github.com/kb-dk/ds-present/releases/tag/ds-present-2.2.5) 2025-01-06
 ### Added